+++ src/repositories/InventoryRepository.h
@@ -17,9 +17,11 @@ public:
     explicit InventoryRepository(QSqlDatabase db, AuditRepository *audit = nullptr,
                                  QObject *parent = nullptr);
 
-    // Registra movimiento con before/after explícitos (no toca stock)
-    bool record(const QString &sku, const QString &productName, const QString &type, int qty,
-                int before, int after, const QString &reason, const QString &user);
+    // Registra movimiento con before/after explícitos (no toca stock).
+    // Fase 2: cantidades decimales.
+    bool record(const QString &sku, const QString &productName, const QString &type,
+                double qty, double before, double after, const QString &reason,
+                const QString &user);
 
     QList<InventoryMovement> movements(int limit = 20) const;
     QList<InventoryMovement> movementsBySku(const QString &sku) const;
